Allow underscores in language and database params
This was blocking queries for languages that have non-letter characters in the language abbreviation, such as nds-nl. However, the database names actually use underscores in place of hyphens for such languages (with exceptions), so we'll be converting hyphens to underscores before making queries.

See https://quarry.wmflabs.org/query/4031 to look up the database name of a particular wiki, to find exceptions to the typical pattern.